Add MCommands and Firmwares Enum
Add MCommands:
Insert all values of M commands in the Enum list, based
on RepRap GCode documentation.
Create the method toString() to return what the command do.
Implemented only the ones that insterested. Jump the CNC ones,
and commands that any firmware support.
Add Firmwares:
Add this Enum to control the type of the firmware in use.
The idea is to create a list or map, to control if the
command is suported by the firmware.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>